==> Preprocesamiento/0-Data_concat.ipynb <==
## Codigo para concatenar todos los csv que ha generado el scrapper

import csv   
import pandas as pd
import numpy as np

data1 =  pd.read_csv('film_actors.csv',  sep='\t', engine='python', header= 0)
data2 =  pd.read_csv('film_actors_1.csv',  sep='\t', engine='python', header= 0)
data3 =  pd.read_csv('film_actors_2.csv',  sep='\t', engine='python', header= 0)
data4 =  pd.read_csv('film_actors_3.csv',  sep='\t', engine='python', header= 0)
data5 =  pd.read_csv('film_actors_4.csv',  sep='\t', engine='python', header= 0)
data6 =  pd.read_csv('film_actors_5.csv',  sep='\t', engine='python', header= 0)
data7 =  pd.read_csv('film_actors_6.csv',  sep='\t', engine='python', header= 0)
data8 =  pd.read_csv('film_actors_7.csv',  sep='\t', engine='c', header= 0)
data9 =  pd.read_csv('film_actors_8.csv',  sep='\t', engine='python', header= 0)
data10 =  pd.read_csv('film_actors_9.csv',  sep='\t', engine='python', header= 0)
data11 =  pd.read_csv('film_actors_likes_def.csv',  sep='\t', engine='python', header= 0)

data13 =  pd.read_csv('film_actors_likes_def_2.csv',  sep='\t', engine='python', header= 0)
data14 =  pd.read_csv('film_actors_likes_def_3.csv',  sep='\t', engine='python', header= 0)
data15 =  pd.read_csv('film_actors_likes_def_4.csv',  sep='\t', engine='python', header= 0)
data16 =  pd.read_csv('film_actors_likes_def_5.csv',  sep='\t', engine='python', header= 0)
data17 =  pd.read_csv('film_actors_likes_def_6.csv',  sep='\t', engine='python', header= 0)
data18 =  pd.read_csv('film_actors_likes_def_7.csv',  sep='\t', engine='python', header= 0)
data19 =  pd.read_csv('film_actors_likes_def_8.csv',  sep='\t', engine='python', header= 0)
data20 =  pd.read_csv('film_actors_likes_def_9.csv',  sep='\t', engine='python', header= 0)
data21 =  pd.read_csv('film_actors_likes_def_10.csv',  sep='\t', engine='python', header= 0)
data22 =  pd.read_csv('film_actors_likes_def_11.csv',  sep='\t', engine='python', header= 0)
data23 =  pd.read_csv('film_actors_likes_def_12.csv',  sep='\t', engine='python', header= 0)
data24 =  pd.read_csv('film_actors_likes_def_13.csv',  sep='\t', engine='python', header= 0)


data8=data8.dropna()

data = pd.concat([data1,data2,data3,data4,data5,data6,data7,data8,data9,data10,data11,data13,data14,data15,data16,data17,\
                  data18,data19,data20,data21,data22,data23,data24])

data.reset_index(inplace = True)

data = data.drop(['index'],axis = 1)

data = data.drop(['ACTOR1_FACEBOOK_LIKES', 'ACTOR1_INSTAGRAM_LIKES','ACTOR1_TWITTER_LIKES','ACTOR2_FACEBOOK_LIKES','ACTOR2_INSTAGRAM_LIKES','ACTOR2_TWITTER_LIKES','ACTOR3_FACEBOOK_LIKES','ACTOR3_INSTAGRAM_LIKES','ACTOR3_TWITTER_LIKES'], axis=1)

data= data.drop_duplicates()

data.to_csv('FINAL_IMBD_ACTORS.csv') 

## Código para juntar con los datos que ya han sido descargados

from google.colab import drive
drive.mount('/content/drive')

data = pd.read_csv('/content/drive/My Drive/TFM/data/imdb_data.csv',header=0,decimal=',')

data2 = pd.read_csv('/content/drive/My Drive/TFM/data/FINAL_IMBD_ACTORS.csv',header=0,decimal=',')

data3 = pd.merge(data, data2, on='CODE')

data3.to_csv('COMPLETE_IMBD_DATA.csv')
