==> Preprocesamiento/2-Filtered_per_images.ipynb <==
# Libraries

import os
import pandas as pd

from google.colab import drive
drive.mount('/content/drive')

filtered_data = pd.read_csv("../data/FILTERED_FINAL_IMBD_DATA.csv", sep=',', engine='python', header=0,index_col=0)
code_list_filtered_data = list(filtered_data['CODE'])
import pandas as pd

string = '.png'

filtered_data['CODE'] = (pd.Series(code_list_filtered_data) + string).tolist()


filtered_ListImages = (pd.Series(code_list_filtered_data) + string).tolist()
len(filtered_ListImages)

filtered_data

list_img_allow = os.listdir('../data/img')

dirImages = set(list_img_allow)
filtered_ListImages = set(filtered_ListImages)
films_in_comon = dirImages.intersection(filtered_ListImages)

films_in_comon =list(films_in_comon)

films_in_comon = pd.DataFrame(films_in_comon,columns=['CODE'])

films_in_comon

filtered_df_with_img = films_in_comon.set_index('CODE').join(filtered_data.set_index('CODE'))


filtered_df_with_img.head(3)

filtered_df_with_img.to_csv('filtered_df_with_imgNames.csv')
