==> Preprocesamiento/4-Tabular_Preprocess_with_actors.ipynb <==
import pandas as pd

data_tab= pd.read_csv('/content/drive/My Drive/Máster KSchool/TFM/data/datos_antiguos/COMPLETE_IMBD_DATA.csv', sep=',', header= 0)
data_tab.columns = ['Unnamed: 0', 'CODE', 'TITLE', 'DESCRIPTION', 'DURATION', 'COLOR',
       'YEAR', 'DIRECTOR', 'BUDGET', 'CWG', 'RATING', 'VOTES',
       'METASCORE RATING', 'GENRES', 'LANGUAGE', 'COUNTRY', 'CONTENT_RATING',
       'KEYWORDS', 'WRITERS', 'OPENING WEEKEND', 'GENRES_0', 'GENRES_1',
       'GENRES_2', 'LANGUAGE_0', 'KEYWORDS_0', 'KEYWORDS_1', 'KEYWORDS_2',
       'COUNTRY_0', 'WRITERS_0', 'KEYWORDS_DESCRIPTION', 'ACTOR_0',
       'ACTOR_1', 'ACTOR_2']
data_tab['ACTOR_2'] = data_tab['ACTOR_2'].astype(str)
data_tab = data_tab.drop(columns=['Unnamed: 0'])
data_tab.head(1)

data_tab['ACTOR_0'] = data_tab['ACTOR_0'].apply(lambda x: x.replace(' ',''))
data_tab['ACTOR_1'] = data_tab['ACTOR_0'].apply(lambda x: x.replace(' ',''))
data_tab['ACTOR_2'] = data_tab['ACTOR_2'].apply(lambda x: x.replace(' ',''))

data_tab.head(3)

print(len(data_tab.DIRECTOR.unique()),
len(data_tab.ACTOR_2.unique()))

def listWithTheBest(data_tab,columnName,number=3):

  data =  data_tab[columnName].value_counts()>number
  data = data.reset_index()

  final_list = data.loc[data[columnName]==True]
  final_list = final_list['index'].values
  return final_list

def new_valuesInDataframe(data_tab,listWithTeBest,columnName,nuevoNombre):
  for i, row in data_tab[columnName].iteritems():
    if row not in listWithTeBest:
      data_tab.at[i,columnName]=nuevoNombre
  return data_tab

best_writers = listWithTheBest(data_tab,'ACTOR_0',3)
data_tab = new_valuesInDataframe(data_tab,best_writers,'ACTOR_0',"PepitoDoe")

best_directors = listWithTheBest(data_tab,'ACTOR_1',3)
data_tab = new_valuesInDataframe(data_tab,best_directors,'ACTOR_1',"PepitaDoe")

best_directors = listWithTheBest(data_tab,'ACTOR_2',3)
data_tab = new_valuesInDataframe(data_tab,best_directors,'ACTOR_2',"JosefinaDoe")

print(len(data_tab.ACTOR_0.unique()),
len(data_tab.ACTOR_1.unique()),
len(data_tab.ACTOR_2.unique()))

data_tab.to_csv("imdb_data.csv",sep=",")
