==> Preprocesamiento/5-get_votes.ipynb <==
Como en el proceso del preprocesamiento se modificaron el tipo de datos de la columna votes nos vemos obligados a volver a extraerlo del dataframe original

import pandas as pd
import numpy as np

datos_with_votes = pd.read_csv("/content/drive/My Drive/Máster KSchool/TFM/data/datos_antiguos/FILTERED_FINAL_IMBD_DATA.csv",sep=',',decimal=',',header=0)
datos = pd.read_csv("/content/drive/My Drive/Máster KSchool/TFM/data/imdb_data.csv",sep=',',decimal=',',header=0)
datos = datos.drop(columns=['Unnamed: 0', 'VOTES'])

datos_with_votes['VOTES']  = datos_with_votes['VOTES'].astype(str)
datos_with_votes['VOTES'] = datos_with_votes['VOTES'].apply(lambda x: x.replace('.',''))
datos_with_votes['VOTES']  = datos_with_votes['VOTES'].astype(int)

'''aux = datos_with_votes[datos_with_votes['CODE']=='tt3280916']
aux.describe()'''

display(datos_with_votes[datos_with_votes['CODE']=='tt3280916'])

datos_with_votes = datos_with_votes[['CODE','VOTES']]

datos_with_votes.head()

datos = datos.merge(datos_with_votes,left_on='CODE', right_on='CODE')

datos[datos['CODE']=='tt3280916']

datos.describe()






datos[datos['CODE']=='tt3280916']


datos.to_csv("imdb_data.csv")

aux = pd.read_csv("imdb_data.csv",sep=',')

aux[aux['CODE']=='tt3280916']
aux['VOTES']  = aux['VOTES'].astype(int)

aux.dtypes

aux.describe()
