==> Preprocesamiento/6-Filter_per_votes.ipynb <==
import pandas as pd
import numpy as np

from google.colab import drive
drive.mount('/content/drive')

datos = pd.read_csv("/content/drive/My Drive/TFM/data/imdb_data.csv",sep=',',decimal=',',header=0)

min_votes = 125
datos = datos.loc[datos['VOTES'] > min_votes]

datos.to_csv("imdb_data.csv")
